refactor tests to use fixtures
