+++ tests/test_mssql.py
@@ -1,52 +1,4 @@
-import sys, types
-if "mysql" not in sys.modules:
-    dummy_mysql = types.ModuleType("mysql")
-    dummy_mysql.connector = types.SimpleNamespace(connect=lambda **k: None)
-    sys.modules["mysql"] = dummy_mysql
-    sys.modules["mysql.connector"] = dummy_mysql.connector
-if "dotenv" not in sys.modules:
-    mod=types.ModuleType("dotenv")
-    mod.load_dotenv=lambda *a, **k: None
-    sys.modules["dotenv"] = mod
-if "sqlalchemy" not in sys.modules:
-    sa_mod = types.ModuleType("sqlalchemy")
-    sa_mod.create_engine = lambda *a, **k: None
-    sa_mod.MetaData = lambda *a, **k: None
-    pool_mod = types.ModuleType("pool")
-    pool_mod.NullPool = object
-    sa_mod.pool = pool_mod
-    engine_mod = types.ModuleType("engine")
-    engine_mod.Engine = object
-    engine_mod.Connection = object
-    engine_mod.URL = types.SimpleNamespace(create=lambda *a, **k: None)
-    sa_mod.engine = engine_mod
-    sys.modules["sqlalchemy"] = sa_mod
-    sys.modules["sqlalchemy.engine"] = engine_mod
-    sys.modules["sqlalchemy.pool"] = pool_mod
-if "pyodbc" not in sys.modules:
-    sys.modules["pyodbc"] = types.SimpleNamespace(connect=lambda *a, **k: None)
-if "pydantic" not in sys.modules:
-    pd_mod = types.ModuleType("pydantic")
-    class _BaseSettings:
-        def __init__(self, **values):
-            for k, v in values.items():
-                setattr(self, k, v)
-    pd_mod.BaseSettings = _BaseSettings
-    pd_mod.DirectoryPath = str
-    pd_mod.Field = lambda *a, **k: None
-    class _SecretStr(str):
-        def get_secret_value(self):
-            return str(self)
-    pd_mod.SecretStr = _SecretStr
-    def _validator(*a, **k):
-        def dec(func):
-            return func
-        return dec
-    pd_mod.validator = _validator
-    sys.modules["pydantic"] = pd_mod
-    ps_mod = types.ModuleType("pydantic_settings")
-    ps_mod.BaseSettings = _BaseSettings
-    sys.modules["pydantic_settings"] = ps_mod
+import sys
 import db.connections as connections
 from config import settings
