+++ tests/test_run_etl.py
@@ -6,82 +6,6 @@ import importlib.util
 from pathlib import Path
 import queue
 
-# Stub optional dependencies used by run_etl when importing
-if "tkinter" not in sys.modules:
-    tk = types.ModuleType("tkinter")
-    tk.Tk = object
-    tk.Label = object
-    tk.Entry = object
-    tk.Button = object
-    tk.Checkbutton = object
-    tk.Frame = object
-    tk.BooleanVar = object
-    tk.StringVar = object
-    tk.scrolledtext = types.SimpleNamespace(ScrolledText=object)
-    tk.messagebox = types.SimpleNamespace(showerror=lambda *a, **k: None,
-                                          showinfo=lambda *a, **k: None)
-    tk.filedialog = types.SimpleNamespace(askdirectory=lambda *a, **k: "")
-    sys.modules["tkinter"] = tk
-    sys.modules["tkinter.messagebox"] = tk.messagebox
-    sys.modules["tkinter.scrolledtext"] = tk.scrolledtext
-    sys.modules["tkinter.filedialog"] = tk.filedialog
-
-if "pyodbc" not in sys.modules:
-    class _DummyError(Exception):
-        pass
-    sys.modules["pyodbc"] = types.SimpleNamespace(Error=_DummyError, connect=lambda *a, **k: None)
-
-if "pandas" not in sys.modules:
-    sys.modules["pandas"] = types.ModuleType("pandas")
-
-if "sqlalchemy" not in sys.modules:
-    sa_mod = types.ModuleType("sqlalchemy")
-    types_mod = types.SimpleNamespace(Text=lambda *a, **k: None)
-    sa_mod.types = types_mod
-    sa_mod.MetaData = lambda *a, **k: None
-    engine_mod = types.ModuleType("engine")
-    engine_mod.Engine = object
-    engine_mod.Connection = object
-    sa_mod.engine = engine_mod
-    sys.modules["sqlalchemy"] = sa_mod
-    sys.modules["sqlalchemy.types"] = types_mod
-    sys.modules["sqlalchemy.engine"] = engine_mod
-
-if "dotenv" not in sys.modules:
-    mod = types.ModuleType("dotenv")
-    mod.load_dotenv = lambda *a, **k: None
-    sys.modules["dotenv"] = mod
-
-if "pydantic" not in sys.modules:
-    pd_mod = types.ModuleType("pydantic")
-    class _BaseSettings:
-        def __init__(self, **values):
-            for k, v in values.items():
-                setattr(self, k, v)
-    pd_mod.BaseSettings = _BaseSettings
-    pd_mod.DirectoryPath = str
-    pd_mod.Field = lambda *a, **k: None
-    class _SecretStr(str):
-        def get_secret_value(self):
-            return str(self)
-    pd_mod.SecretStr = _SecretStr
-    def _validator(*a, **k):
-        def dec(func):
-            return func
-        return dec
-    pd_mod.validator = _validator
-    sys.modules["pydantic"] = pd_mod
-    ps_mod = types.ModuleType("pydantic_settings")
-    ps_mod.BaseSettings = _BaseSettings
-    sys.modules["pydantic_settings"] = ps_mod
-
-if "tqdm" not in sys.modules:
-    dummy = types.ModuleType("tqdm")
-    def _tqdm(iterable, **kwargs):
-        for item in iterable:
-            yield item
-    dummy.tqdm = _tqdm
-    sys.modules["tqdm"] = dummy
 
 
 def _import_run_etl_from_repo(tmp_cwd):
